feat(graphic): add ShaderResourceBinder class
Introduces the initial scaffolding for the `ShaderResourceBinder` class.
This class will be responsible for managing the binding of various
graphics resources, such as constant buffers, textures, and samplers, to
shader stages.

This initial commit provides the basic class definition and file
structure, which will be built upon in subsequent changes.