refactor(core): Mark GetRawData as const
The GetRawData method does not modify the state of the Image object, as it only returns a const pointer to the internal pixel data.

Adding the `const` qualifier to the method signature enforces this contract at the compiler level. This improves const-correctness and allows the method to be safely called on constant Image instances or references.

Signed-off-by: CaIzii <2470159730@qq.com>