feat(graphic): Add upload method stubs to D12DepthTexture
Implement the `UploadToGPU` and `GetUploadDestinationState` virtual methods for the `D12DepthTexture` class.

These methods are required to conform to the base resource interface for a generalized upload system. The current implementations are stubs that will trigger a fatal error, as uploading initial data to a depth texture is not yet supported.

This change satisfies the interface contract and allows the project to compile.

Signed-off-by: CaIzii <2470159730@qq.com>